docs(expo): update Expo getting started for multi-store API
Restructure the getting started guide to reflect the new StoreRegistry-based API:
- Add new "Configure the store" section with store.ts snippet
- Rename provider section to "Set up the store registry"
- Update snippets to use useAppStore custom hook pattern
- Update queries section to use store.useQuery() method
